increase precision, centralize scale

+++ src/face_tracking.py
@@ -46,7 +46,7 @@ def face_delta(face_detector, image: cv2.Mat) -> Optional[tuple[int, int]]:
     """
     Gets the nose position relative to the center of the camera frame.
     """
-    center = (int(image.shape[1] / 2), int(image.shape[0] / 2))
+    center = (image.shape[1] / 2, image.shape[0] / 2)
     detections = face_detections(face_detector, image)
     if detections:
         closest: mediapipe.framework.formats.detection_pb2.Detection = max(
